Move initClusterFrame around
