+++ packages/core/src/features/electron/renderer/frame-routing-id.injectable.ts
@@ -3,7 +3,7 @@
  * Licensed under MIT License. See LICENSE in root directory for more information.
  */
 import { getInjectable } from "@ogre-tools/injectable";
-import webFrameInjectable from "./web-frame/web-frame.injectable";
+import webFrameInjectable from "./web-frame.injectable";
 
 const frameRoutingIdInjectable = getInjectable({
   id: "frame-routing-id",
